Update the first few lectures to provide better structure
This was sorely needed, now that I only have lectures and other people
conduct the programming lab sessions. Split the first lecture in half to
separate the "Into to programming" from "Intro to C++" (and provide a
template for "Intro to OCaml", "Intro to Python", etc). Expand the
second lecture a bit and cut the content that is only for the "hardcore"
track -- it will reappear later anyway.